COMP:Fix warning on VS 64bit. Don't why gcc 4.3.2 didn't catch this one on a 64bit machine with -Wconversion on.

+++ Source/kwsys/RegularExpression.cxx
@@ -348,7 +348,7 @@ static int strcspn ();
 bool RegularExpression::compile (const char* exp) {
     register const char* scan;
     register const char* longest;
-    register unsigned long len;
+    register size_t len;
              int         flags;
 
     if (exp == 0) {
